Move preferences-related classes into a separate package
