+++ app/src/main/AndroidManifest.xml
@@ -75,14 +75,14 @@
             android:excludeFromRecents="true"
             android:theme="@style/ProxyTheme" />
 
-        <activity android:name="fr.gaulupeau.apps.Poche.ui.SettingsActivityNew">
+        <activity android:name="fr.gaulupeau.apps.Poche.ui.preferences.SettingsActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MANAGE_NETWORK_USAGE" />
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
         </activity>
 
-        <activity android:name="fr.gaulupeau.apps.Poche.ui.wizard.ConnectionWizardActivity"/>
+        <activity android:name="fr.gaulupeau.apps.Poche.ui.preferences.ConnectionWizardActivity"/>
 
         <receiver android:name="android.support.v4.media.session.MediaButtonReceiver">
             <intent-filter>
